+++ scripts/test_catalyst.sh
@@ -53,8 +53,8 @@ args=(
   "ARCHS=x86_64" "VALID_ARCHS=x86_64" "SUPPORTS_MACCATALYST=YES"
   # Run on macOS.
   "-sdk" "macosx" "-destination platform=\"OS X\"" "TARGETED_DEVICE_FAMILY=2"
-  # Disable signing.
-  "CODE_SIGN_IDENTITY=-" "CODE_SIGNING_REQUIRED=NO" "CODE_SIGNING_ALLOWED=NO"
+  # Use ad-hoc signing for macOS tests.
+  "CODE_SIGN_IDENTITY=-"
   # GHA is still running 10.15.
   "MACOSX_DEPLOYMENT_TARGET=10.15"
 )