+++ FirebaseFunctions/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- [fixed] Replaced unsafe uses of `os_unfair_lock` (#14548).
+
 # 11.10.0
 - [added] Streaming callable functions are now supported.